+++ src/main/java/com/demo/service/StudentServices.java
@@ -36,7 +36,11 @@ public class StudentServices {
 	
 	public Student getById(Long id)
 	{
-		return studentDao.getById(id);
+		Student student = studentDao.getById(id);
+		if(student==null){
+			throw new NullPointerException("Student Id is not found -" +id);
+		}
+		return student;
 	}
 	
 	//update student
@@ -49,6 +53,9 @@ public class StudentServices {
 	
 	public void deleteStudent(Long id)
 	{
+		if(id==null){
+			throw new NullPointerException("No such student with id-"+id);
+		}
 		studentDao.deleteStudent(id);
 	}
